removed single variant and moved styles to baseStyles
