+++ design/atoms/components/input/theme/index.ts
@@ -1,11 +1,9 @@
 import baseStyle from './input.base'
-import variants from './input.variants'
 import sizes from './input.sizes'
 import defaultProps from './input.default'
 
 export default {
   baseStyle,
-  variants,
   sizes,
   defaultProps,
 };
