+++ design/atoms/components/input/theme/input.default.ts
@@ -1,4 +1,3 @@
 export default {
-  variant: 'primary',
   size: 'base',
 }
